Revert "Split out LibraryBar"
This reverts commit bffa25f915b355169f404928eef3807852d7ac35.
