+++ src/components/Sections/Library/LibraryContainer.js
@@ -9,22 +9,22 @@ import { withNamespaces } from 'react-i18next';
 import { Button, Container, Grid, Header, Input, Ref, Segment } from 'semantic-ui-react';
 import Headroom from 'react-headroom';
 
-import { isEmpty } from '../../../helpers/utils';
+import { formatError, isEmpty } from '../../../helpers/utils';
 import { actions as assetsActions, selectors as assets } from '../../../redux/modules/assets';
 import { actions as sourceActions, selectors as sources } from '../../../redux/modules/sources';
 import { selectors as settings } from '../../../redux/modules/settings';
 import * as shapes from '../../shapes';
-import { getSourceErrorSplash } from '../../shared/WipErr/WipErr';
+import { ErrorSplash, FrownSplash } from '../../shared/Splash/Splash';
 import Helmets from '../../shared/Helmets';
-import { isTaas } from '../../shared/PDF/PDF';
 import LibraryContentContainer from './LibraryContentContainer';
 import TOC, { getIndex } from './TOC';
-import LibraryBar from './LibraryBar';
-import { getLanguageDirection } from '../../../helpers/i18n-utils';
+import LibrarySettings from './LibrarySettings';
+import Share from './Share';
+import { getLanguageDirection, isLanguageRtl } from '../../../helpers/i18n-utils';
 import { DeviceInfoContext } from '../../../helpers/app-contexts';
 import { getQuery } from '../../../helpers/url';
 import { SCROLL_SEARCH_ID } from '../../../helpers/consts';
-
+import { isTaas } from '../../shared/PDF/PDF';
 
 const waitForRenderElement = async (attempts = 0) => {
   if (attempts > 10) return Promise.reject();
@@ -322,6 +322,9 @@ class LibraryContainer extends Component {
     }
   };
 
+  print = () => {
+    window.print();
+  };
 
   matchString = (parentId, t) => {
     const { NotToFilter } = this.props;
@@ -358,6 +361,10 @@ class LibraryContainer extends Component {
     return path;
   };
 
+  static getErrContent = (err, t) => (err.response && err.response.status === 404)
+    ? <FrownSplash text={t('messages.source-content-not-found')} />
+    : <ErrorSplash text={t('messages.server-error')} subtext={formatError(err)} />;
+
   getContent = () => {
     const { sourceId, indexMap, language, contentLanguage, t, history } = this.props;
     const index                                                         = isEmpty(sourceId) ? {} : indexMap[sourceId];
@@ -366,7 +373,7 @@ class LibraryContainer extends Component {
     let content;
 
     if (err) {
-      content = getSourceErrorSplash(err, t);
+      content = LibraryContainer.getErrContent(err, t);
     } else {
       const downloadAllowed = this.context.deviceInfo.os.name !== 'iOS';
       content               = (
@@ -457,6 +464,9 @@ class LibraryContainer extends Component {
     const fullPath    = LibraryContainer.getFullPath(sourceId, getPathByID);
     const parentId    = this.properParentId(fullPath);
     const matchString = this.matchString(parentId, t);
+
+    const isRtl    = isLanguageRtl(language);
+    const position = isRtl ? 'left' : 'right';
     const active   = !this.context.isMobileDevice || tocIsActive;
 
     return (
@@ -494,13 +504,14 @@ class LibraryContainer extends Component {
                   </Grid.Column>
                   <Grid.Column mobile={16} tablet={16} computer={12} className="source__content-header">
                     <div className="source__header-title">{this.header(sourceId, parentId)}</div>
-                    <LibraryBar
-                      handleSettings={this.handleSettings}
-                      handleIsReadable={this.handleIsReadable}
-                      handleTocIsActive={this.handleTocIsActive}
-                      isReadable={isReadable}
-                      fontSize={fontSize}>
-                    </LibraryBar>
+                    <div className="source__header-toolbar">
+                      <Button compact size="small" className="mobile-hidden" icon="print" onClick={this.print} />
+                      <div id="download-button" />
+                      <LibrarySettings fontSize={fontSize} handleSettings={this.handleSettings} />
+                      <Button compact size="small" icon={isReadable ? 'compress' : 'expand'} onClick={this.handleIsReadable} />
+                      <Button compact size="small" className="computer-hidden large-screen-hidden widescreen-hidden" icon="list layout" onClick={this.handleTocIsActive} />
+                      <Share position={position} />
+                    </div>
                   </Grid.Column>
                 </Grid.Row>
               </Grid>
